+++ lib/wasabi/core_ext/object.rb
@@ -1,0 +1,14 @@
+module Wasabi
+  module CoreExt
+    module Object
+
+      # Returns +true+ if the Object is nil, false or empty. Implementation from ActiveSupport.
+      def blank?
+        respond_to?(:empty?) ? empty? : !self
+      end unless method_defined?(:blank?)
+
+    end
+  end
+end
+
+Object.send :include, Wasabi::CoreExt::Object